Unified metadata panel between videos and images
