+++ frontend/components/dashboard/image-detail-modal.tsx
@@ -1,16 +1,14 @@
 import React, { useState, useRef, useCallback, useEffect } from 'react'
 import { motion, AnimatePresence } from 'framer-motion'
-import { X, MessageCircle, Send, Calendar, User, ZoomIn, ZoomOut, Crop, Edit2, Save, X as XIcon, ChevronLeft, ChevronRight, Play, Pause, FileText, Image as ImageIcon, HardDrive, PanelLeftClose, PanelLeftOpen, Clock } from 'lucide-react'
+import { X, MessageCircle, Send, ZoomIn, ZoomOut, Crop, ChevronLeft, ChevronRight, Play, Pause, PanelLeftClose, PanelLeftOpen } from 'lucide-react'
 import { useRouter } from 'next/navigation'
-import { useMediaComments, useCreateComment, useAuth, useUpdateMedia } from '@/lib/api-hooks'
-import { getMediaUrlFromMedia } from '@/lib/api'
+import { useMediaComments, useCreateComment, useAuth } from '@/lib/api-hooks'
 import { getSmartMediaUrl } from '@/lib/media-utils'
-import { LazyMedia } from '../lazy-media'
-import { TagInput } from '../tag-input'
 import { Media, Comment } from '@/lib/types'
 import { mapMediaData } from '@/lib/media-utils'
 import { useSlideshow } from '@/lib/hooks/use-slideshow'
 import { ModalEventCatcher } from '@/components/common/modal-event-catcher'
+import { MediaMetadataPanel } from '@/components/common/media-metadata-panel'
 
 interface ImageDetailModalProps {
   media: Media | null
@@ -41,36 +39,14 @@ export function ImageDetailModal({ media, onClose, onMediaUpdate, updateMedia, a
   const [isCropping, setIsCropping] = useState(false)
   const [cropStart, setCropStart] = useState({ x: 0, y: 0 })
   const [activeCrop, setActiveCrop] = useState<{ x: number, y: number, width: number, height: number } | null>(null)
-  const [isEditing, setIsEditing] = useState(false)
-  const [editTitle, setEditTitle] = useState('')
-  const [editDescription, setEditDescription] = useState('')
-  const [editTags, setEditTags] = useState<string[]>([])
-  const [localTitle, setLocalTitle] = useState(media?.altText || '')
-  const [localDescription, setLocalDescription] = useState(media?.caption || '')
-  const [localTags, setLocalTags] = useState<string[]>(media?.tags || [])
   const modalRef = useRef<HTMLDivElement>(null)
-  
-  // Unsaved changes tracking
-  const [hasUnsavedChanges, setHasUnsavedChanges] = useState(false)
-  const [showUnsavedChangesDialog, setShowUnsavedChangesDialog] = useState(false)
-  const [pendingNavigation, setPendingNavigation] = useState<(() => void) | null>(null)
   
 
   
-  // Update local state when media prop changes (for navigation)
+
+  
+  // Update view state when media prop changes (for navigation)
   useEffect(() => {
-    setLocalTitle(media?.altText || '')
-    setLocalDescription(media?.caption || '')
-    setLocalTags(media?.tags || [])
-    // Reset edit state when navigating to a new media
-    setIsEditing(false)
-    setEditTitle('')
-    setEditDescription('')
-    setEditTags([])
-    // Reset unsaved changes tracking
-    setHasUnsavedChanges(false)
-    setShowUnsavedChangesDialog(false)
-    setPendingNavigation(null)
     // Reset view state when navigating to a new media
     // We'll set zoom to 1 initially, then adjust it in handleImageLoad
     console.log('setting zoom to 1')
@@ -81,7 +57,7 @@ export function ImageDetailModal({ media, onClose, onMediaUpdate, updateMedia, a
     
     // We'll defer optimal scaling to handleImageLoad
     // This ensures we have the actual image dimensions
-  }, [media?.id, media?.altText, media?.caption, media?.tags])
+  }, [media?.id])
   const containerRef = useRef<HTMLDivElement>(null)
   
   // UI state
@@ -91,7 +67,6 @@ export function ImageDetailModal({ media, onClose, onMediaUpdate, updateMedia, a
   
   const { data: commentsData, isLoading: commentsLoading } = useMediaComments(media.id || '')
   const createCommentMutation = useCreateComment()
-  const updateMediaMutation = useUpdateMedia()
   const { data: authData } = useAuth()
   
   // Check if current user is the media owner
@@ -111,50 +86,29 @@ export function ImageDetailModal({ media, onClose, onMediaUpdate, updateMedia, a
   const hasNext = canNavigate && currentIndex !== -1 && currentIndex < allMedia.length - 1
   const hasPrev = canNavigate && currentIndex !== -1 && currentIndex > 0
 
-  // Navigation with unsaved changes handling
-  const handleNavigationWithUnsavedChanges = useCallback((navigationFunction: () => void) => {
-    if (hasUnsavedChanges) {
-      setShowUnsavedChangesDialog(true)
-      setPendingNavigation(() => navigationFunction)
-    } else {
-      navigationFunction()
-    }
-  }, [hasUnsavedChanges])
 
-  const handleConfirmNavigation = useCallback(() => {
-    setShowUnsavedChangesDialog(false)
-    if (pendingNavigation) {
-      pendingNavigation()
-      setPendingNavigation(null)
-    }
-  }, [pendingNavigation])
-
-  const handleCancelNavigation = useCallback(() => {
-    setShowUnsavedChangesDialog(false)
-    setPendingNavigation(null)
-  }, [])
 
   const handleNext = useCallback(() => {
     if (!canNavigate || !hasNext) return
     const nextIndex = currentIndex + 1
-    handleNavigationWithUnsavedChanges(() => onNavigate(allMedia[nextIndex]))
-  }, [canNavigate, hasNext, currentIndex, onNavigate, allMedia, handleNavigationWithUnsavedChanges])
+    onNavigate(allMedia[nextIndex])
+  }, [canNavigate, hasNext, currentIndex, onNavigate, allMedia])
 
   const handlePrev = useCallback(() => {
     if (!canNavigate || !hasPrev) return
     const prevIndex = currentIndex - 1
-    handleNavigationWithUnsavedChanges(() => onNavigate(allMedia[prevIndex]))
-  }, [canNavigate, hasPrev, currentIndex, onNavigate, allMedia, handleNavigationWithUnsavedChanges])
+    onNavigate(allMedia[prevIndex])
+  }, [canNavigate, hasPrev, currentIndex, onNavigate, allMedia])
 
   const handleFirst = useCallback(() => {
     if (!canNavigate) return
-    handleNavigationWithUnsavedChanges(() => onNavigate(allMedia[0]))
-  }, [canNavigate, onNavigate, allMedia, handleNavigationWithUnsavedChanges])
+    onNavigate(allMedia[0])
+  }, [canNavigate, onNavigate, allMedia])
 
   const handleLast = useCallback(() => {
     if (!canNavigate) return
-    handleNavigationWithUnsavedChanges(() => onNavigate(allMedia[allMedia.length - 1]))
-  }, [canNavigate, onNavigate, allMedia, handleNavigationWithUnsavedChanges])
+    onNavigate(allMedia[allMedia.length - 1])
+  }, [canNavigate, onNavigate, allMedia])
 
   // Handle image load for slideshow and initial scaling
   const handleImageLoad = useCallback((e: React.SyntheticEvent<HTMLImageElement>) => {
@@ -192,41 +146,7 @@ export function ImageDetailModal({ media, onClose, onMediaUpdate, updateMedia, a
     }
   }, [slideshow, isPanelMinimized])
 
-  // Slideshow speed control functions
-  const formatSlideshowSpeed = (speed: number) => {
-    if (speed < 1000) {
-      return `${speed}ms`
-    } else {
-      return `${Math.round(speed / 1000)}s`
-    }
-  }
 
-  // Logarithmic scale for slideshow speed
-  // Convert slider value (0-100) to speed (3000-60000ms)
-  const sliderToSpeed = (sliderValue: number) => {
-    // Use logarithmic scale: 3s to 60s
-    const minSpeed = 3000 // 3 seconds
-    const maxSpeed = 60000 // 60 seconds
-    const logMin = Math.log(minSpeed)
-    const logMax = Math.log(maxSpeed)
-    const logValue = logMin + (logMax - logMin) * (sliderValue / 100)
-    return Math.round(Math.exp(logValue))
-  }
-
-  // Convert speed to slider value
-  const speedToSlider = (speed: number) => {
-    const minSpeed = 3000
-    const maxSpeed = 60000
-    const logMin = Math.log(minSpeed)
-    const logMax = Math.log(maxSpeed)
-    const logSpeed = Math.log(speed)
-    return Math.round(((logSpeed - logMin) / (logMax - logMin)) * 100)
-  }
-
-  const handleSlideshowSpeedChange = (sliderValue: number) => {
-    const newSpeed = sliderToSpeed(sliderValue)
-    slideshow.updateSlideshowSpeed(newSpeed)
-  }
 
   const toggleCropMode = () => {
     console.log('toggleCropMode called, current isCropMode:', isCropMode)
@@ -292,17 +212,9 @@ export function ImageDetailModal({ media, onClose, onMediaUpdate, updateMedia, a
   }
 
   const handleClose = useCallback(() => {
-    if (hasUnsavedChanges) {
-      setShowUnsavedChangesDialog(true)
-      setPendingNavigation(() => () => {
-        resetView()
-        onClose()
-      })
-    } else {
-      resetView()
-      onClose()
-    }
-  }, [hasUnsavedChanges, resetView, onClose])
+    resetView()
+    onClose()
+  }, [resetView, onClose])
 
   // Keyboard navigation
   useEffect(() => {
@@ -377,80 +289,7 @@ export function ImageDetailModal({ media, onClose, onMediaUpdate, updateMedia, a
     }
   }
 
-  // Track changes in edit fields
-  const handleEditTitleChange = (value: string) => {
-    setEditTitle(value)
-    // Only check for changes if we're in editing mode
-    if (isEditing) {
-      const hasChanges = value !== localTitle || editDescription !== localDescription || JSON.stringify(editTags) !== JSON.stringify(localTags)
-      setHasUnsavedChanges(hasChanges)
-    }
-  }
 
-  const handleEditDescriptionChange = (value: string) => {
-    setEditDescription(value)
-    // Only check for changes if we're in editing mode
-    if (isEditing) {
-      setHasUnsavedChanges(editTitle !== localTitle || value !== localDescription || JSON.stringify(editTags) !== JSON.stringify(localTags))
-    }
-  }
-
-  const handleEditTagsChange = (tags: string[]) => {
-    setEditTags(tags)
-    // Only check for changes if we're in editing mode
-    if (isEditing) {
-      setHasUnsavedChanges(editTitle !== localTitle || editDescription !== localDescription || JSON.stringify(tags) !== JSON.stringify(localTags))
-    }
-  }
-
-  const handleStartEdit = () => {
-    setEditTitle(localTitle)
-    setEditDescription(localDescription)
-    setEditTags([...localTags])
-    setIsEditing(true)
-    setHasUnsavedChanges(false)
-  }
-
-  const handleCancelEdit = () => {
-    setIsEditing(false)
-    setEditTitle('')
-    setEditDescription('')
-    setEditTags([])
-    setHasUnsavedChanges(false)
-  }
-
-  const handleSaveEdit = async () => {
-    try {
-      if (!media.id) return
-      await updateMediaMutation.mutateAsync({
-        mediaId: media.id,
-        title: editTitle,
-        description: editDescription,
-        tags: editTags
-      })
-      
-      setLocalTitle(editTitle)
-      setLocalDescription(editDescription)
-      setLocalTags([...editTags])
-      setIsEditing(false)
-      setHasUnsavedChanges(false)
-      
-      // Update the media prop if updateMedia function is provided
-      if (updateMedia && media.id) {
-        updateMedia(media.id, {
-          altText: editTitle,
-          caption: editDescription,
-          tags: editTags
-        })
-      }
-      
-      if (onMediaUpdate) {
-        onMediaUpdate()
-      }
-    } catch (error) {
-      console.error('Failed to update media:', error)
-    }
-  }
 
 
 
@@ -885,134 +724,23 @@ export function ImageDetailModal({ media, onClose, onMediaUpdate, updateMedia, a
               <MessageCircle className="h-4 w-4 text-gray-400" />
             </div>
           ) : (
-            <div className="p-4 border-b border-gray-200">
-              <div className="flex items-start justify-between mb-2">
-                {isEditing ? (
-                  <div className="flex-1 space-y-3">
-                    <input
-                      type="text"
-                      value={editTitle}
-                      onChange={(e) => handleEditTitleChange(e.target.value)}
-                      placeholder="Image title..."
-                      className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent text-sm font-medium"
-                    />
-                    <textarea
-                      value={editDescription}
-                      onChange={(e) => handleEditDescriptionChange(e.target.value)}
-                      placeholder="Image description..."
-                      rows={3}
-                      className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent text-sm resize-none"
-                    />
-                    <TagInput
-                      tags={editTags}
-                      onTagsChange={handleEditTagsChange}
-                      placeholder="Add tags..."
-                      className="w-full"
-                    />
-                    <div className="flex space-x-2">
-                      <button
-                        onClick={handleSaveEdit}
-                        disabled={updateMediaMutation.isPending}
-                        className="px-4 py-2 bg-blue-600 text-white rounded-lg hover:bg-blue-700 transition-colors duration-200 text-sm"
-                      >
-                        {updateMediaMutation.isPending ? 'Saving...' : 'Save'}
-                      </button>
-                      <button
-                        onClick={handleCancelEdit}
-                        className="px-4 py-2 bg-gray-300 text-gray-700 rounded-lg hover:bg-gray-400 transition-colors duration-200 text-sm"
-                      >
-                        Cancel
-                      </button>
-                    </div>
-                  </div>
-                ) : (
-                  <div className="flex-1">
-                    <h3 className="text-lg font-semibold text-gray-900 mb-1">
-                      {localTitle || 'Untitled Image'}
-                    </h3>
-                    {localDescription && (
-                      <p className="text-gray-600 text-sm mb-2">{localDescription}</p>
-                    )}
-                    {localTags.length > 0 && (
-                      <div className="flex flex-wrap gap-1 mb-2">
-                        {localTags.map((tag, index) => (
-                          <span
-                            key={index}
-                            className="px-2 py-1 bg-blue-100 text-blue-800 text-xs rounded-full"
-                          >
-                            {tag}
-                          </span>
-                        ))}
-                      </div>
-                    )}
-                    <div className="flex items-center space-x-4 text-xs text-gray-500">
-                      <span className="flex items-center">
-                        <Calendar className="h-3 w-3 mr-1" />
-                        {formatDate(media.createdAt)}
-                      </span>
-                      <span className="flex items-center">
-                        <User className="h-3 w-3 mr-1" />
-                        <button
-                          onClick={() => handleAuthorClick(media.authorId, null)}
-                          className="hover:text-blue-600 transition-colors duration-200"
-                        >
-                          Unknown User
-                        </button>
-                      </span>
-                      {media.originalFilename && (
-                        <span className="flex items-center">
-                          <FileText className="h-3 w-3 mr-1" />
-                          {media.originalFilename}
-                        </span>
-                      )}
-                    </div>
-                  </div>
-                )}
-                
-                {!isEditing && isOwner && (
-                  <button
-                    onClick={handleStartEdit}
-                    className="p-2 text-gray-400 hover:text-gray-600 hover:bg-gray-100 rounded-lg transition-colors duration-200"
-                    title="Edit Image"
-                  >
-                    <Edit2 className="h-4 w-4" />
-                  </button>
-                )}
-              </div>
-            </div>
+            <MediaMetadataPanel
+              media={media}
+              isOwner={isOwner}
+              onMediaUpdate={onMediaUpdate}
+              updateMedia={updateMedia}
+              slideshow={{
+                isSlideshowActive: slideshow.isSlideshowActive,
+                currentSlideshowSpeed: slideshow.currentSlideshowSpeed,
+                updateSlideshowSpeed: slideshow.updateSlideshowSpeed
+              }}
+              canNavigate={!!canNavigate}
+              allMedia={allMedia}
+              mediaType="IMAGE"
+            />
           )}
 
-          {/* Slideshow Speed Control */}
-                        {slideshow.isSlideshowActive && canNavigate && allMedia && allMedia.length > 1 && (
-            <div className="p-4 border-b border-gray-200 bg-blue-50">
-              <div className="flex items-center space-x-2 mb-3">
-                <Clock className="h-4 w-4 text-blue-600" />
-                <h4 className="text-sm font-medium text-gray-900">Slideshow Speed</h4>
-              </div>
-              <div className="space-y-2">
-                <div className="flex items-center justify-between">
-                  <span className="text-xs text-gray-600">3s</span>
-                  <span className="text-xs text-gray-600">60s</span>
-                </div>
-                <input
-                  type="range"
-                  min="0"
-                  max="100"
-                  value={speedToSlider(slideshow.currentSlideshowSpeed)}
-                  onChange={(e) => handleSlideshowSpeedChange(parseInt(e.target.value))}
-                  className="w-full h-2 bg-gray-200 rounded-lg appearance-none cursor-pointer slider"
-                  style={{
-                    background: `linear-gradient(to right, #3b82f6 0%, #3b82f6 ${speedToSlider(slideshow.currentSlideshowSpeed)}%, #e5e7eb ${speedToSlider(slideshow.currentSlideshowSpeed)}%, #e5e7eb 100%)`
-                  }}
-                />
-                <div className="text-center">
-                  <span className="text-sm font-medium text-blue-600">
-                    {formatSlideshowSpeed(slideshow.currentSlideshowSpeed)}
-                  </span>
-                </div>
-              </div>
-            </div>
-          )}
+
 
           {/* Comments */}
           <div className="flex-1 overflow-y-auto">
@@ -1090,48 +818,7 @@ export function ImageDetailModal({ media, onClose, onMediaUpdate, updateMedia, a
         </motion.div>
       </motion.div>
 
-            {/* Unsaved Changes Dialog */}
-      <AnimatePresence>
-        {showUnsavedChangesDialog && (
-          <div className="fixed inset-0 z-[9999] flex items-center justify-center bg-black bg-opacity-50">
-            <div className="bg-white rounded-lg p-6 max-w-md w-full mx-4 shadow-xl">
-              <h3 className="text-lg font-semibold text-gray-900 mb-4">
-                Unsaved Changes
-              </h3>
-              <p className="text-gray-600 mb-6">
-                You have unsaved changes to this media. Do you want to save them before continuing?
-              </p>
-              <div className="flex space-x-3">
-                <button
-                  onClick={handleCancelNavigation}
-                  className="flex-1 px-4 py-2 border border-gray-300 text-gray-700 rounded-lg hover:bg-gray-50 transition-colors duration-200"
-                >
-                  Cancel
-                </button>
-                <button
-                  onClick={handleConfirmNavigation}
-                  className="flex-1 px-4 py-2 bg-gray-600 text-white rounded-lg hover:bg-gray-700 transition-colors duration-200"
-                >
-                  Discard Changes
-                </button>
-                <button
-                  onClick={async () => {
-                    await handleSaveEdit()
-                    if (pendingNavigation) {
-                      pendingNavigation()
-                      setPendingNavigation(null)
-                    }
-                    setShowUnsavedChangesDialog(false)
-                  }}
-                  className="flex-1 px-4 py-2 bg-blue-600 text-white rounded-lg hover:bg-blue-700 transition-colors duration-200"
-                >
-                  Save & Continue
-                </button>
-              </div>
-            </div>
-          </div>
-        )}
-      </AnimatePresence>
+
     </AnimatePresence>
     </ModalEventCatcher>
   )
